On_mention and cogs implemented
